refactor(proxy): split into modular structure and add WebSocket proxy
- Split monolithic proxy.rs into proxy/ module (mod, external,
  inject, response, rewrite, websocket)
- Add tokio-tungstenite for WebSocket reverse proxy support
